Creación de módulos
Creación de módulos- función enmascaramiento(prototipo)
