+++ Code/ProjectParams.pro
@@ -1,3 +1,7 @@
 QT += core gui
 CONFIG += console c++17
-SOURCES += main.cpp
+SOURCES += main.cpp \
+    funciones_definidas.cpp
+
+HEADERS += \
+    funciones.h
